outbound: don't double-wrap replay bodies
In PR #1020, the retry code was changed so that bodies are only wrapped
in a `ReplayBody` when the request will be retried. The intention was to
remove the `replay::layer` pushed to the stack before the retry layer,
so that bodies are only wrapped when retries are enabled for the route,
and not for all routes in a profile.

However, I somehow managed to not remove the `replay::layer` from the
stack (possibly due to a messed up rebase). Oops. This meant that all
bodies for endpoints with service profiles were wrapped, and when the
route has retries enabled, it's wrapped twice. When wrapped twice, this
means that we potentially buffer the body data *twice*.